Binary Search Bug
While(lo is less than hi) should be
While (intLo is less than intHi) as the while loop is an indicator of the boundaries of the array i.e. while the lowest and highest boundries of an array do not cross, do xyz.
But the if statements do the comparisons of actual values IN the array, to decide where the lowest and highest value of the array goes.